We *should* be able to use smaller types for the flood-fill decode heap

+++ src/lib/cimb_translator/FloodDecodePositions.h
@@ -12,8 +12,8 @@ class FloodDecodePositions
 {
 public:
 	using iter = std::tuple<unsigned, CellPositions::coordinate, CellDrift>;
-	using decode_instructions = std::pair<CellDrift, unsigned>; // drift, best_prio
-	using decode_prio = std::tuple<unsigned, unsigned>; // index, prio
+	using decode_instructions = std::pair<CellDrift, uint8_t>; // drift, best_prio
+	using decode_prio = std::tuple<uint16_t, uint8_t>; // index, prio
 
 	class PrioCompare
 	{
